+++ lib/src/manager/state/grid_state.dart
@@ -24,6 +24,10 @@ abstract class IGridState {
 
   PlutoOnRowSecondaryTapEventCallback? get onRowSecondaryTap;
 
+  PlutoOnRowEnterEventCallback? get onRowEnter;
+
+  PlutoOnRowExitEventCallback? get onRowExit;
+
   PlutoOnRowsMovedEventCallback? get onRowsMoved;
 
   PlutoOnColumnsMovedEventCallback? get onColumnsMoved;
@@ -33,6 +37,8 @@ abstract class IGridState {
   CreateHeaderCallBack? get createHeader;
 
   CreateFooterCallBack? get createFooter;
+
+  PlutoSelectDateCallBack? get selectDateCallback;
 
   PlutoGridLocaleText get localeText;