+++ lib/src/pluto_grid_configuration.dart
@@ -198,6 +198,7 @@ class PlutoGridStyleConfig {
     this.enableCellBorderVertical = true,
     this.enableCellBorderHorizontal = true,
     this.enableRowColorAnimation = false,
+    this.enableRowHoverColor = false,
     this.gridBackgroundColor = Colors.white,
     this.rowColor = Colors.white,
     this.oddRowColor,
@@ -206,6 +207,7 @@ class PlutoGridStyleConfig {
     Color? columnCheckedColor,
     Color? cellCheckedColor,
     this.rowCheckedColor = const Color(0x11757575),
+    this.rowHoveredColor = const Color(0xFFB1B3B7),
     this.cellColorInEditState = Colors.white,
     this.cellColorInReadOnlyState = const Color(0xFFDBDBDC),
     this.cellColorGroupedRow,
@@ -251,6 +253,8 @@ class PlutoGridStyleConfig {
     this.rowGroupEmptyIcon = Icons.noise_control_off,
     this.gridBorderRadius = BorderRadius.zero,
     this.gridPopupBorderRadius = BorderRadius.zero,
+    this.gridPadding = PlutoGridSettings.gridPadding,
+    this.gridBorderWidth = PlutoGridSettings.gridBorderWidth,
     this.filterHeaderColor,
     this.filterHeaderIconColor,
   })  : columnCheckedColor = (columnCheckedColor ?? activatedColor),
@@ -267,6 +271,7 @@ class PlutoGridStyleConfig {
     this.enableCellBorderVertical = true,
     this.enableCellBorderHorizontal = true,
     this.enableRowColorAnimation = false,
+    this.enableRowHoverColor = false,
     this.gridBackgroundColor = const Color(0xFF111111),
     this.rowColor = const Color(0xFF111111),
     this.oddRowColor,
@@ -275,6 +280,7 @@ class PlutoGridStyleConfig {
     Color? columnCheckedColor,
     Color? cellCheckedColor,
     this.rowCheckedColor = const Color(0x11202020),
+    this.rowHoveredColor = const Color(0xFF3D3D3D),
     this.cellColorInEditState = const Color(0xFF666666),
     this.cellColorInReadOnlyState = const Color(0xFF222222),
     this.cellColorGroupedRow,
@@ -320,6 +326,8 @@ class PlutoGridStyleConfig {
     this.rowGroupEmptyIcon = Icons.noise_control_off,
     this.gridBorderRadius = BorderRadius.zero,
     this.gridPopupBorderRadius = BorderRadius.zero,
+    this.gridPadding = PlutoGridSettings.gridPadding,
+    this.gridBorderWidth = PlutoGridSettings.gridBorderWidth,
     this.filterHeaderColor,
     this.filterHeaderIconColor,
   })  : columnCheckedColor = (columnCheckedColor ?? activatedColor),
@@ -347,6 +355,14 @@ class PlutoGridStyleConfig {
   /// Animation of background color transition of rows,
   /// such as when the current row or rows are dragged.
   final bool enableRowColorAnimation;
+
+  /// Hover effect on rows.
+  /// If true, the background color of the row changes to [rowHoveredColor]
+  /// when the mouse hovers over it.
+  /// If false, the background color of the row does not change and
+  /// the background color of the row is the same as [rowColor].
+  /// [rowHoveredColor] is therefore not used.
+  final bool enableRowHoverColor;
 
   final Color gridBackgroundColor;
 
@@ -378,6 +394,9 @@ class PlutoGridStyleConfig {
 
   /// Checked Color for the row. (Checked rows)
   final Color rowCheckedColor;
+
+  /// Hovered Color. (Currently hovered row)
+  final Color rowHoveredColor;
 
   /// Cell color in edit state. (only current cell)
   final Color cellColorInEditState;
@@ -492,6 +511,12 @@ class PlutoGridStyleConfig {
   /// Apply border radius to popup opened inside [PlutoGrid].
   final BorderRadiusGeometry gridPopupBorderRadius;
 
+  /// Defaults to [PlutoGridSettings.gridPadding]
+  final double gridPadding;
+
+  /// Defaults to [PlutoGridSettings.gridBorderWidth]
+  final double gridBorderWidth;
+
   /// Set color of filter popup header
   final Color? filterHeaderColor;
 
@@ -545,6 +570,8 @@ class PlutoGridStyleConfig {
     IconData? rowGroupEmptyIcon,
     BorderRadiusGeometry? gridBorderRadius,
     BorderRadiusGeometry? gridPopupBorderRadius,
+    double? gridPadding,
+    double? gridBorderWidth,
   }) {
     return PlutoGridStyleConfig(
         enableGridBorderShadow:
@@ -613,6 +640,8 @@ class PlutoGridStyleConfig {
         gridBorderRadius: gridBorderRadius ?? this.gridBorderRadius,
         gridPopupBorderRadius:
             gridPopupBorderRadius ?? this.gridPopupBorderRadius,
+        gridPadding: gridPadding ?? this.gridPadding,
+        gridBorderWidth: gridBorderWidth ?? this.gridBorderWidth,
         filterHeaderColor: filterHeaderColor ?? filterHeaderColor,
         filterHeaderIconColor: filterHeaderIconColor ?? filterHeaderIconColor);
   }
@@ -668,7 +697,9 @@ class PlutoGridStyleConfig {
             rowGroupCollapsedIcon == other.rowGroupCollapsedIcon &&
             rowGroupEmptyIcon == other.rowGroupEmptyIcon &&
             gridBorderRadius == other.gridBorderRadius &&
-            gridPopupBorderRadius == other.gridPopupBorderRadius;
+            gridPopupBorderRadius == other.gridPopupBorderRadius &&
+            gridPadding == other.gridPadding &&
+            gridBorderWidth == other.gridBorderWidth;
   }
 
   @override
@@ -719,6 +750,8 @@ class PlutoGridStyleConfig {
         rowGroupEmptyIcon,
         gridBorderRadius,
         gridPopupBorderRadius,
+        gridPadding,
+        gridBorderWidth,
       ]);
 }