+++ lib/src/ui/pluto_body_rows.dart
@@ -101,10 +101,12 @@ class PlutoBodyRowsState extends PlutoStateWithChange<PlutoBodyRows> {
             scrollDirection: Axis.vertical,
             physics: const ClampingScrollPhysics(),
             itemCount: _rows.length,
-            itemExtent: stateManager.rowTotalHeight,
+            itemExtent: stateManager.rowWrapper != null
+                ? null
+                : stateManager.rowTotalHeight,
             addRepaintBoundaries: false,
             itemBuilder: (ctx, i) {
-              return PlutoBaseRow(
+              Widget w = PlutoBaseRow(
                 key: ValueKey('body_row_${_rows[i].key}'),
                 rowIdx: i,
                 row: _rows[i],
@@ -112,6 +114,10 @@ class PlutoBodyRowsState extends PlutoStateWithChange<PlutoBodyRows> {
                 stateManager: stateManager,
                 visibilityLayout: true,
               );
+
+              return stateManager.rowWrapper != null
+                  ? stateManager.rowWrapper(w)
+                  : w;
             },
           ),
         ),