+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [8.4.1] - 2024. 05. 15
+upgrade for flutter 3.22.0
+
+## [8.4.0] - 2024. 04. 01
+
+
+* Added onRowEnter, onRowExit callbacks to react on. @coruscant187
+* Added logic to change background color of row if hovered. @coruscant187 doonfrs/pluto_grid_plus#29
+
 ## [8.3.1] - 2024. 03. 19
 * upgrade packages to latest major version